+++ noweda/__init__.py
@@ -1,6 +1,7 @@
 from .io import read, read_chunked
+from .large import LargeDataset
 from .accessor import NowEDAAccessor
 
 from ._version import __version__
 
-__all__ = ["read", "read_chunked", "__version__"]
+__all__ = ["read", "read_chunked", "LargeDataset", "__version__"]